# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NyumbaLink API
  version: "1.0.0"
  description: |
    ## Kenya's Property Marketplace REST API

    **Base URL:** `https://nyumbalink.co.ke/api`

    ### Authentication
    All protected endpoints require a JWT Bearer token obtained from `/auth/login`.
    Include it as: `Authorization: Bearer <access_token>`

    ### Rate Limiting
    | Scope           | Limit             |
    |-----------------|-------------------|
    | Global          | 200 req/day, 50/hr |
    | Auth endpoints  | 10 req/min        |
    | File uploads    | 20 req/hr         |

    ### Pagination
    List endpoints accept:
    | Param     | Type    | Default | Description              |
    |-----------|---------|---------|--------------------------|
    | page      | integer | 1       | Page number (1-indexed)  |
    | per_page  | integer | 12      | Items per page (max 50)  |
    | sort_by   | string  | created_at | Column to sort by     |
    | sort_dir  | string  | desc    | `asc` or `desc`          |

  contact:
    name: NyumbaLink Support
    email: support@nyumbalink.co.ke
    url: https://nyumbalink.co.ke

  license:
    name: Proprietary
    url: https://nyumbalink.co.ke/terms
servers:
  - url: https://nyumbalink.co.ke/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# ──────────────────────────────────────────────────────────────────────
#  Reusable schemas
# ──────────────────────────────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:     { type: integer, example: 248 }
        page:      { type: integer, example: 1 }
        per_page:  { type: integer, example: 12 }
        pages:     { type: integer, example: 21 }
        has_prev:  { type: boolean, example: false }
        has_next:  { type: boolean, example: true }
        sort_by:   { type: string,  example: created_at }
        sort_dir:  { type: string,  enum: [asc, desc], example: desc }
        prev_url:  { type: string,  nullable: true }
        next_url:  { type: string,  nullable: true }
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string, example: "Resource not found" }
        details: { type: object, description: "Field-level validation errors" }

    # ── User ──────────────────────────────────────────────────────────
...
    PropertyImage:
      type: object
      properties:
        id:         { type: string, format: uuid }
        image_url:  { type: string, format: uri }
        is_primary: { type: boolean }
        sort_order: { type: integer }
    Property:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        title:               { type: string, example: "3 Bed Apartment – Kilimani" }
        description:         { type: string, nullable: true }
        category:            { type: string, enum: [rent, short_stay, plot_sale] }
        price:               { type: number, format: float, example: 45000 }
        price_period:        { type: string, enum: [monthly, per_night, yearly, total], nullable: true }
        location:            { type: string, example: "Kilimani, Nairobi" }
        county:              { type: string, example: "Nairobi" }
        town:                { type: string, example: "Kilimani" }
        latitude:            { type: number, nullable: true }
        longitude:           { type: number, nullable: true }
        bedrooms:            { type: integer, nullable: true }
        bathrooms:           { type: integer, nullable: true }
        plot_size:           { type: number, nullable: true }
        plot_size_unit:      { type: string, nullable: true }
        floor_area:          { type: number, nullable: true }
        amenities:           { type: array, items: { type: string } }
        owner_id:            { type: string, format: uuid }
        verification_status: { type: string, enum: [pending, verified, rejected] }
        status:              { type: string, enum: [active, inactive, sold, rented] }
        featured:            { type: boolean }
        views_count:         { type: integer }
        images:              { type: array, items: { $ref: '#/components/schemas/PropertyImage' } }
        primary_image:       { $ref: '#/components/schemas/PropertyImage', nullable: true }
        avg_rating:          { type: number, nullable: true }
        created_at:          { type: string, format: date-time }
        updated_at:          { type: string, format: date-time }
    PropertyCreate:
      type: object
      required: [title, category, price, location]
      properties:
        title:        { type: string, minLength: 5, maxLength: 500 }
        description:  { type: string, maxLength: 5000 }
        category:     { type: string, enum: [rent, short_stay, plot_sale] }
        price:        { type: number, minimum: 0 }
        price_period: { type: string, enum: [monthly, per_night, yearly, total] }
        location:     { type: string }
        county:       { type: string }
        town:         { type: string }
        latitude:     { type: number }
        longitude:    { type: number }
        bedrooms:     { type: integer, minimum: 0 }
        bathrooms:    { type: integer, minimum: 0 }
        plot_size:    { type: number, minimum: 0 }
        floor_area:   { type: number, minimum: 0 }
        amenities:    { type: array, items: { type: string } }

    # ── Inquiry ───────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /properties:
    get:
      summary: List published properties (with filters + pagination)
      operationId: listProperties
      tags: [Properties]
      parameters:
        - { name: page,        in: query, schema: { type: integer, default: 1 } }
        - { name: per_page,    in: query, schema: { type: integer, default: 12, maximum: 50 } }
        - { name: sort_by,     in: query, schema: { type: string, enum: [price, created_at, views_count, bedrooms] } }
        - { name: sort_dir,    in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - { name: category,    in: query, schema: { type: string, enum: [rent, short_stay, plot_sale] } }
        - { name: county,      in: query, schema: { type: string }, example: Nairobi }
        - { name: town,        in: query, schema: { type: string }, example: Kilimani }
        - { name: min_price,   in: query, schema: { type: number } }
        - { name: max_price,   in: query, schema: { type: number } }
        - { name: bedrooms,    in: query, schema: { type: integer } }
        - { name: min_plot,    in: query, schema: { type: number } }
        - { name: q,           in: query, schema: { type: string }, description: Keyword search }
        - { name: featured,    in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Paginated property list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: '#/components/schemas/Property' } }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }

    post:
      summary: Create a new property listing (owners only)
      operationId: createProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyCreate' }
      responses:
        "201":
          description: Property created (status=pending verification)
          content:
            application/json:
              schema:
                type: object
                properties:
                  property: { $ref: '#/components/schemas/Property' }
                  message:  { type: string }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /properties/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      summary: Get property details (public)
      operationId: getProperty
      tags: [Properties]
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Property' }
        "404": { $ref: '#/components/responses/NotFound' }

    put:
      summary: Update property (owner or admin)
      operationId: updateProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PropertyCreate' }
      responses:
        "200": { description: Updated property }
        "403": { $ref: '#/components/responses/Forbidden' }

    delete:
      summary: Delete property (owner or admin)
      operationId: deleteProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      responses:
        "200": { description: Deleted }
        "403": { $ref: '#/components/responses/Forbidden' }
  /properties/{id}/images:
    post:
      summary: Upload images for a property
      operationId: uploadImages
      tags: [Properties]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "201": { description: Images uploaded }
  /properties/my/listings:
    get:
      summary: List my own property listings
      operationId: myListings
      tags: [Properties]
      security:
        - BearerAuth: []
      parameters:
        - { name: page,     in: query, schema: { type: integer } }
        - { name: per_page, in: query, schema: { type: integer } }
      responses:
        "200": { description: Owner's listings }
  /properties/{id}/save:
    post:
      summary: Toggle save/unsave a property
      operationId: saveProperty
      tags: [Properties]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Saved or unsaved
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved:   { type: boolean }
                  message: { type: string }

  # ── Search ────────────────────────────────────────────────────────
...